+++ internal/gitops/gitops.go
@@ -7,6 +7,7 @@ import (
 	"github.com/mih-kopylov/bulker/internal/shell"
 	"github.com/spf13/afero"
 	"os"
+	"regexp"
 	"strings"
 )
 
@@ -26,6 +27,32 @@ const (
 	StatusDirty
 	StatusMissing
 	StatusError
+)
+
+type GitMode string
+
+func (g *GitMode) String() string {
+	return string(*g)
+}
+
+func (g *GitMode) Set(v string) error {
+	switch v {
+	case string(GitModeAll), string(GitModeLocal), string(GitModeRemote):
+		*g = GitMode(v)
+		return nil
+	default:
+		return fmt.Errorf("must be one of '%s' '%s' '%s'", GitModeAll, GitModeLocal, GitModeRemote)
+	}
+}
+
+func (g *GitMode) Type() string {
+	return "GitMode"
+}
+
+const (
+	GitModeAll    GitMode = "all"
+	GitModeLocal  GitMode = "local"
+	GitModeRemote GitMode = "remote"
 )
 
 func CloneRepo(fs afero.Fs, repo *model.Repo, recreate bool) (CloneResult, error) {
@@ -115,6 +142,66 @@ func Status(fs afero.Fs, repo *model.Repo) (StatusResult, error) {
 	}
 
 	return StatusDirty, nil
+}
+
+func GetBranches(fs afero.Fs, repo *model.Repo, mode GitMode, pattern string) ([]Branch, error) {
+	err := checkRepoExists(fs, repo)
+	if err != nil {
+		return nil, err
+	}
+
+	reg, err := regexp.Compile("^" + pattern + "$")
+	if err != nil {
+		return nil, err
+	}
+
+	outputString, err := shell.RunCommand(repo.Path, "git", "branch", "-a", "--format=%(refname)")
+	if err != nil {
+		return nil, fmt.Errorf("failed to get branches: %w", err)
+	}
+
+	branches, err := parseBranches(outputString)
+	if err != nil {
+		return nil, err
+	}
+
+	var result []Branch
+	for _, branch := range branches {
+		if mode == GitModeLocal && !branch.IsLocal() {
+			continue
+		}
+
+		if mode == GitModeRemote && branch.IsLocal() {
+			continue
+		}
+
+		matched := reg.MatchString(branch.Name)
+		if !matched {
+			continue
+		}
+
+		result = append(result, branch)
+	}
+
+	return result, nil
+
+}
+
+func parseBranches(consoleOutputString string) ([]Branch, error) {
+	var result []Branch
+	for _, outputBranchName := range strings.Fields(consoleOutputString) {
+		branch, err := parseBranch(outputBranchName)
+		if err != nil {
+			return nil, err
+		}
+
+		if branch.Name == Head {
+			continue
+		}
+		result = append(result, *branch)
+	}
+
+	return result, nil
 }
 
 func checkRepoExists(fs afero.Fs, repo *model.Repo) error {